requirements.txt
appdirs==1.4.0
appnope==0.1.0
bleach==1.5.0
decorator==4.0.11
dj-database-url==0.4.1
Django==1.9.7
django-extensions==1.7.6
django-filter==1.0.1
djangorestframework==3.5.4
entrypoints==0.2.2
gunicorn==19.6.0
html5lib==0.9999999
ipdb==0.10.2
ipykernel==4.5.2
ipython==5.2.2
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Markdown==2.6.8
MarkupSafe==0.23
mistune==0.7.3
nbconvert==5.1.1
nbformat==4.2.0
notebook==4.4.1
olefile==0.44
packaging==16.8
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
pyzmq==16.0.2
qtconsole==4.2.1
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
whitenoise==2.0.6
widgetsnbextension==1.2.6
